refactor about current
