+++ src/app/components/TopButton/TopButton.tsx
@@ -1,15 +1,14 @@
 'use client';
 
 import { Button } from '@nextui-org/button';
-import { useDisclosure } from '@nextui-org/react';
+import { useDisclosure } from '@nextui-org/use-disclosure';
 import { FC } from 'react';
 import { EditNoteModal } from '~/components/domains/Note/EditNoteModal';
 import { LoginModal } from '~/components/domains/User/LoginModal';
 import { Icon } from '~/components/uiParts/icons';
-import { useCurrentUser } from '~/hooks/user/useCurrentUser';
+import { User } from '~/domains/User';
 
-export const TopButton: FC = () => {
-  const { data: currentUser } = useCurrentUser();
+export const TopButton: FC<{ currentUser: User }> = ({ currentUser }) => {
   const { isOpen: isOpenPostNoteModal, onOpen: onOpenPostNoteModal, onOpenChange: onOpenChangePostNoteModal } = useDisclosure();
   const { isOpen: isOpenLoginModal, onOpen: onOpenLoginModal, onOpenChange: onOpenChangeLoginModal } = useDisclosure();
