+++ src/app/layout.tsx
@@ -1,6 +1,7 @@
 import { Metadata } from 'next';
 import { Providers } from './providers';
 import '../styles/global.scss';
+import { fetchMe } from './actions/userActions';
 import { WiscroNavbar } from '~/components/layouts/WiscroNavbar';
 import { WiscroFooter } from '~/components/layouts/WiscroFooter';
 
@@ -9,11 +10,13 @@ export const metadata: Metadata = {
   description: 'WiscroはURLをまとめたページを気軽に作れるアプリケーションです。情報の整理や共有にお使いください。',
 };
 
-export default function RootLayout({ children }: { children: React.ReactNode }) {
+export default async function RootLayout({ children }: { children: React.ReactNode }) {
+  const { currentUser } = await fetchMe();
+
   return (
     <html lang="ja" className="light min-h-[100vh] bg-slate-50">
       <body className="min-h-[100vh] flex flex-col">
-        <WiscroNavbar />
+        <WiscroNavbar currentUser={currentUser} />
         <div className="flex-1">
           <Providers>{children}</Providers>
         </div>
