+++ src/components/layouts/LoginRequiredWrapper/LoginRequiredWrapper.tsx
@@ -1,32 +1,24 @@
 'use client';
 
-import { Spinner } from '@nextui-org/react';
 import { useRouter } from 'next/navigation';
 import { FC, ReactNode, useEffect } from 'react';
+import { fetchMe } from '~/app/actions/userActions';
 import { URLS } from '~/constants/urls';
-import { useCurrentUser } from '~/hooks/user/useCurrentUser';
 
 type Props = {
   children: ReactNode;
 };
 
-export const LoginRequiredWrapper: FC<Props> = ({ children }) => {
-  const { data: currentUser, isLoading: isLoadingCurrentUser } = useCurrentUser();
+export const LoginRequiredWrapper: FC<Props> = async ({ children }) => {
+  const { currentUser } = await fetchMe();
   const router = useRouter();
 
   useEffect(() => {
-    if (isLoadingCurrentUser || currentUser) {
+    if (currentUser) {
       return;
     }
     router.push(`${URLS.TOP}?isRedirect=true`);
-  }, [currentUser, isLoadingCurrentUser, router]);
-
-  if (isLoadingCurrentUser)
-    return (
-      <div className="w-100 flex justify-center py-[16px]">
-        <Spinner size="lg" color="secondary" />
-      </div>
-    );
+  }, [currentUser, router]);
 
   if (!currentUser) return null;
