+++ src/components/layouts/WiscroNavbar/WiscroNavbar.tsx
@@ -1,14 +1,13 @@
 'use client';
 
-import React from 'react';
-import { Navbar, NavbarBrand, NavbarContent, NavbarItem, Link, Button, Spinner, Avatar, useDisclosure } from '@nextui-org/react';
-import { useCurrentUser } from '~/hooks/user/useCurrentUser';
+import React, { FC } from 'react';
+import { Navbar, NavbarBrand, NavbarContent, NavbarItem, Link, Button, Avatar, useDisclosure } from '@nextui-org/react';
 import { LoginButton } from '~/components/domains/User/LoginButton';
 import { Icon } from '~/components/uiParts/icons';
 import { EditNoteModal } from '~/components/domains/Note/EditNoteModal';
+import { User } from '~/domains/User';
 
-export const WiscroNavbar = () => {
-  const { data: currentUser, isLoading } = useCurrentUser();
+export const WiscroNavbar: FC<{ currentUser: User }> = ({ currentUser }) => {
   const { isOpen, onOpen, onOpenChange } = useDisclosure();
 
   return (
@@ -20,23 +19,19 @@ export const WiscroNavbar = () => {
           </Link>
         </NavbarBrand>
         <NavbarContent justify="end">
-          {isLoading ? (
-            <Spinner />
-          ) : (
-            <NavbarItem>
-              {currentUser ? (
-                <div className="flex gap-[16px]">
-                  <Button color="primary" className="color-[6366f1]" onClick={onOpen}>
-                    <Icon icon="BOOK" />
-                    作成する
-                  </Button>
-                  <Avatar src={currentUser.profileUrl} isBordered />
-                </div>
-              ) : (
-                <LoginButton />
-              )}
-            </NavbarItem>
-          )}
+          <NavbarItem>
+            {currentUser ? (
+              <div className="flex gap-[16px]">
+                <Button color="primary" className="color-[6366f1]" onClick={onOpen}>
+                  <Icon icon="BOOK" />
+                  作成する
+                </Button>
+                <Avatar src={currentUser.profileUrl} isBordered />
+              </div>
+            ) : (
+              <LoginButton />
+            )}
+          </NavbarItem>
         </NavbarContent>
       </Navbar>
       <EditNoteModal isOpen={isOpen} onOpenChange={onOpenChange} />
