+++ src/constants/apiUrls.ts
@@ -2,3 +2,6 @@
 export const API_NOTE_DETAIL = (id: string) => `/api/notes/${id}`;
 export const API_NOTE_LIST = () => `/api/notes`;
 export const API_NOTE_PAGE_LIST = (noteId: string) => `/api/notes/${noteId}/pages`;
+
+// User
+export const API_ME = () => `/api/me`;
